implemented zproc into pipeline but needs debugging
